refactoring and refresh token endpoint added

+++ app/services/jwtService.ts
@@ -9,11 +9,11 @@ import {
 import { v4 as uuidv4 } from 'uuid';
 import db from '../database';
 
-export default function generateAccessToken(additionalPayload: any, refreshToken: boolean = true) {
+export default function generateAccessToken(additionalPayload: any, deleteRefreshToken: boolean = true, refreshToken: boolean = true) {
     const iat = Math.floor(Date.now() / 1000);
     const exp = iat + JWT_EXPIRY_LIMIT;
 
-    if (additionalPayload.grant_type !== 'client_credentials') {
+    if (deleteRefreshToken && additionalPayload.grant_type !== 'client_credentials') {
         const deleteQuery = `
             DELETE FROM auth.refresh_tokens
             WHERE user_id = $1 AND grant_type = $2
@@ -32,8 +32,9 @@ export default function generateAccessToken(additionalPayload: any, refreshToken
 
     if (!refreshToken) {
         return {
-            accessToken,
-            exp
+            access_token: accessToken,
+            token_type: 'bearer',
+            expires: exp
         }
     }
 
@@ -47,14 +48,16 @@ export default function generateAccessToken(additionalPayload: any, refreshToken
     db.query(insertQuery, [jti, additionalPayload.sub, 'signIn']);
 
     return {
-        accessToken,
-        refershToken: jwt.sign({
+        access_token: accessToken,
+        refresh_token: jwt.sign({
             iss: ISSUER,
             aud: AUDIENCE,
             ...additionalPayload,
             iat,
             jti
         }, REFRESH_TOKEN_SECRET!),
-        exp
+        token_type: 'bearer',
+        expires: exp,
+
     };
 }
